+++ app/src/main/java/com/ping/app/PingApplication.kt
@@ -1,7 +1,7 @@
 package com.ping.app
 
 import android.app.Application
-import com.ping.app.presentation.util.LocationHelper
+import com.ping.app.ui.util.LocationHelper
 
 class PingApplication : Application() {
     override fun onCreate() {
