+++ app/src/main/java/com/ping/app/data/repository/TestRepo.kt
@@ -1,0 +1,5 @@
+package com.ping.app.data.repository
+
+class TestRepo {
+
+}
