--- app/src/main/java/com/ping/app/data/repository/TestRepoImpl.kt
@@ -1,7 +1,0 @@
-package com.ping.app.data.repository
-
-import com.ping.app.domain.repository.TestRepo
-
-class TestRepoImpl: TestRepo {
-
-}
