--- app/src/main/java/com/ping/app/domain/model/Test.kt
@@ -1,3 +1,0 @@
-package com.ping.app.domain.model
-
-data class Test(val test: String)
