--- app/src/main/java/com/ping/app/domain/repository/TestRepo.kt
@@ -1,6 +1,0 @@
-package com.ping.app.domain.repository
-
-interface TestRepo {
-
-
-}
