--- app/src/main/java/com/ping/app/domain/usecase/TestUsecase.kt
@@ -1,8 +1,0 @@
-package com.ping.app.domain.usecase
-
-class TestUsecase {
-    // invoke로만 접근한다
-    operator fun invoke(){
-        return
-    }
-}
