+++ app/src/main/java/com/ping/app/ui/base/BaseFragment.kt
@@ -1,8 +1,5 @@
-package com.ping.app.presentation.base
+package com.ping.app.ui.base
 
-import android.app.Activity
-import android.app.Dialog
-import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -12,13 +9,10 @@ import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModel
-import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.google.android.material.bottomsheet.BottomSheetDialog
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
-abstract class BaseBottomSheetDialogFragment<B : ViewDataBinding, VM: ViewModel>(
+abstract class BaseFragment<B : ViewDataBinding, VM: ViewModel>(
     @LayoutRes private val layoutId: Int,
-) : BottomSheetDialogFragment() {
+) : Fragment() {
     private var _binding: B? = null
     protected val binding get() = _binding!!
     protected abstract val viewModel: VM
@@ -37,20 +31,6 @@ abstract class BaseBottomSheetDialogFragment<B : ViewDataBinding, VM: ViewModel>
         super.onViewCreated(view, savedInstanceState)
         binding.lifecycleOwner = viewLifecycleOwner
         initView(savedInstanceState)
-    }
-    
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        // used to show the bottom sheet dialog
-        dialog?.setOnShowListener { it ->
-            val d = it as BottomSheetDialog
-            val bottomSheet =
-                d.findViewById<View>(com.google.android.material.R.id.design_bottom_sheet)
-            bottomSheet?.let {
-                val behavior = BottomSheetBehavior.from(it)
-                behavior.state = BottomSheetBehavior.STATE_EXPANDED
-            }
-        }
-        return super.onCreateDialog(savedInstanceState)
     }
 
     protected abstract fun initView(savedInstanceState: Bundle?)
