+++ app/src/main/java/com/ping/app/ui/feature/main/MainFragment.kt
@@ -1,11 +1,10 @@
-package com.ping.app.presentation.ui.feature.main
+package com.ping.app.ui.feature.main
 
 import android.os.Bundle
 import androidx.fragment.app.viewModels
-import androidx.navigation.fragment.findNavController
 import com.ping.app.R
 import com.ping.app.databinding.FragmentMainBinding
-import com.ping.app.presentation.base.BaseFragment
+import com.ping.app.ui.base.BaseFragment
 
 
 class MainFragment : BaseFragment<FragmentMainBinding, MainViewModel>(R.layout.fragment_main) {
