+++ app/src/main/res/layout/fragment_ping_map.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.feature.map.PingMapFragment">
+        tools:context=".ui.feature.map.PingMapFragment">
 
         <ProgressBar
             android:id="@+id/map_progress"
